Renamed Enemy to Bomb and added empty abstract class Enemy.cs extending Monobehaviour and extended by Bomb.cs
