+++ Assets/Prefabs/Bomb.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 1812771460209059533}
   - component: {fileID: 1812771460209059633}
   m_Layer: 8
-  m_Name: Enemy
+  m_Name: Bomb
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
